requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.1.2
diff-match-patch==20121119
Django==1.10
django-filter==1.0.4
-e git+https://github.com/django-import-export/django-import-export.git@78b96f2ad261d79a1f2acd13ca5e6d2f41486edd#egg=django_import_export
djangorestframework==3.6.4
djangorestframework-constant-field==0.9.1
enum34==1.1.6
et-xmlfile==1.0.1
ipython==5.0.0
ipython-genutils==0.2.0
jdcal==1.3
MySQL-python==1.2.5
odfpy==1.3.5
openpyxl==2.4.8
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyYAML==3.12
scandir==1.5
selenium==3.8.0
simplegeneric==0.8.1
six==1.10.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
wheel==0.26.0
xlrd==1.1.0
xlwt==1.3.0
